Separate tests requiring rasterio so they can be skipped conditionally
